New ipython notebook with dummy sales data added
The sales data is a dummy (created with numpy.random.randint) because it's not a good idea to upload real sales data. The rest of the data are open-sourced, I used APIs and other techniques to collect them and I organised them accordingly but they're data you can find in the web.